requirements.txt
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
plotly==5.15.0
scikit-learn==1.3.0
matplotlib==3.7.2